few makefile and new waitpid

+++ 3.ybx/3.apue/6.few/10.mysh_return_malloc_pointer.c
@@ -3,6 +3,8 @@
 #include <unistd.h>
 #include <string.h>
 #include <glob.h>
+#include <sys/types.h>
+#include <sys/wait.h>
 
 #define SIZE 1024
 
@@ -48,7 +50,7 @@ int main(int argc, const char *argv[])
 	pid_t pid;
 	int i;
 	char *str = NULL;
-	int size = 0;
+	size_t size = 0;
 	cmd_st *cmd_p;
 
 	while (1) {
